Break out UI components from CreateCredential
