+++ src/containers/_dialog.css
@@ -17,4 +17,5 @@
     padding-right: 1rem;
     position: sticky;
     top: 0;
+    z-index: 99999;
 }
